refactor: PositiveInteger package name 재변경
