+++ src/main/java/step2/StringCalculator.java
@@ -1,4 +1,4 @@
-package positiveinteger;
+package step2;
 
 import java.util.Arrays;
 import java.util.regex.Matcher;
