+++ src/test/java/step2/StringCalculatorTest.java
@@ -1,4 +1,4 @@
-package positiveinteger;
+package step2;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
